[NEAR-119] Refactor: Concert 모델 분리에 따른 임포트 경로 수정
